# KC_API/api/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "KickConnect API",
    "description": "Auto-generated Swagger documentation",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/get-all-members": {
      "get": {
        "tags": [
          "Membership"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/get-member-by-id": {
      "get": {
        "tags": [
          "Membership"
        ],
        "description": "",
        "parameters": [
          {
            "name": "memberId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/update-member": {
      "put": {
        "tags": [
          "Membership"
        ],
        "description": "",
        "parameters": [
          {
            "name": "memberId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "memberPlanId": {
                  "example": "any"
                },
                "homeLocationId": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                },
                "contactName": {
                  "example": "any"
                },
                "contactPhone": {
                  "example": "any"
                },
                "signupDate": {
                  "example": "any"
                },
                "renewalDate": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/add-member": {
      "post": {
        "tags": [
          "Membership"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "accountId": {
                  "example": "any"
                },
                "memberPlanId": {
                  "example": "any"
                },
                "homeLocationId": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "phone": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                },
                "contactName": {
                  "example": "any"
                },
                "contactPhone": {
                  "example": "any"
                },
                "signupDate": {
                  "example": "any"
                },
                "renewalDate": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}